Rewrote guts of exponential decay.
I hit the limit of 16-bit fixed point math...the highest decay factor
decomposed into linear decay.

So I gutted it.

And rewrote it informed by the DC and ADSR blocks.  It now uses a linear
ramp, that gets squared and applied to the input signal.  Taking
advantage of the fact that X^2 looks somewhat exponential on the
interval 0..1.

It doesn't sound half bad...